♻️  zb: object_server::interface mod now a hierarchy
